+++ Runtime/Scripts/GameFlow/AutoBinder/BindAttribute.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace Dave6.CharacterKit.GameFlow.AutoBinder
+{
+    [AttributeUsage(AttributeTargets.Field)]
+    public class BindAttribute : Attribute {}
+}