+++ Runtime/Scripts/Handler/Sensor/RaycastSensor.cs
@@ -3,10 +3,8 @@ using UnityEngine;
 namespace Dave6.CharacterKit.Sensor
 {
     /// <summary>
-    /// 리펙토링 예정
-    /// Config | 어떻게 쏠 것인가
-    /// Resolver | 어디서 어느 방향으로 쏠 것인가
-    /// Executor | 실제 Physics 호출
+    /// Cast 방향을 보관하고
+    /// 상태를 저장하는 구조
     /// </summary>
     public class RaycastSensor
     {
@@ -60,42 +58,5 @@ namespace Dave6.CharacterKit.Sensor
                 _ => Vector3.one
             };
         }
-    }
-
-    /// <summary>
-    /// 위랑 다른점
-    /// direction 방향이 갱신 안됨
-    /// </summary>
-    public class RaycastSensor2
-    {
-        public float CastLength
-        {
-            get => _Config.Length;
-            set => _Config.Length = value;
-        }
-        public LayerMask Layermask
-        {
-            get => _Config.LayerMask;
-            set => _Config.LayerMask = value;
-        }
-
-        readonly CastConfig _Config = new();
-        readonly CastResolver _Resolver = new();
-
-        Vector3 _Direction;
-        RaycastHit _HitInfo;
-
-        public RaycastSensor2(Transform root) => _Resolver.Root = root;
-
-        public bool Cast() => PhysicsCaster.Cast(_Resolver.ResolveOrigin(), _Direction, _Config, out _HitInfo);
-        public bool HasDetecteHit() => _HitInfo.collider != null;
-        public float GetDistance() => _HitInfo.distance;
-        public Vector3 GetNormal() => _HitInfo.normal;
-        public Vector3 GetPosition() => _HitInfo.point;
-        public Collider GetCollider() => _HitInfo.collider;
-        public void SetRadius(float value) => _Config.Radius = value;
-        public void SetCastOrigin(Vector3 worldPos) => _Resolver.LocalOrigin = _Resolver.Root.InverseTransformPoint(worldPos);
-        public void SetCastDirection(Vector3 direction) => _Direction = direction.normalized;
-    
     }
 }